cmake: translations: some minor improvements
