+++ CMakeLists.txt
@@ -86,6 +86,7 @@ set(QV2RAY_SOURCES
   src/common/LogHighlighter.cpp
   src/common/QJsonModel.cpp
   src/common/QvHelpers.cpp
+  src/common/QvTranslator.cpp
   src/components/autolaunch/QvAutoLaunch.cpp
   src/components/geosite/QvGeositeReader.cpp
   src/components/icmping/win/ICMPPinger.cpp
